Move the runner endpoint to its own package. The goal is to allow comp/process to call these functions
